+++ app/models/MovieInfo.scala
@@ -15,12 +15,12 @@ case class MovieInfo(
                  url: String,
                )
 
-object JsonFormats {
-
-  import play.api.libs.json.Json
-  import reactivemongo.play.json._
-  import reactivemongo.play.json.collection.JSONCollection
-
-  implicit val movieInfoFormat: OFormat[MovieInfo] = Json.format[MovieInfo]
-  implicit val futureReleaseInfoFormat: OFormat[FutureReleaseInfo] = Json.format[FutureReleaseInfo]
-}
+//object JsonFormats {
+//
+//  import play.api.libs.json.Json
+//  import reactivemongo.play.json._
+//  import reactivemongo.play.json.collection.JSONCollection
+//
+//  implicit val movieInfoFormat: OFormat[MovieInfo] = Json.format[MovieInfo]
+//  implicit val futureReleaseInfoFormat: OFormat[FutureReleaseInfo] = Json.format[FutureReleaseInfo]
+//}
